REF: consolidate constraint modifier code
