+++ simplesat/constraints/__init__.py
@@ -1,3 +1,5 @@
 from .package_parser import PrettyPackageStringParser
 from .requirement import Requirement
-from .requirement_transformation import transform_install_requires
+from .constraint_modifiers import (
+    ConstraintModifiers, transform_install_requires
+)
